backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.5.0
python-dotenv>=1.0.0
faker>=18.9.0
django-filter>=23.2
django-storages>=1.13.2
psycopg2-binary>=2.9.6
boto3==1.34.34
djangorestframework-simplejwt>=5.5.0 